Edits to last pull
Splitting function blocks into smaller functions. Other small changes.
