+++ src/examples/example.py
@@ -1,24 +1,24 @@
 # pylint: disable=E1120
 import click
-from examples.acceleration_readings import read_accelerometers
+from examples.acceleration_readings import (read_accelerometers,live_read_accelerometers)
 from examples.aligning_readings import align_acceleration_readings
-from examples.run_pyoma import (
-    run_oma_and_plot,
-    run_oma_and_publish,
-    run_oma_and_print,
-    run_oma_and_publish_loop,
+from examples.run_sysid import (
+    run_sysid_and_plot,
+    run_sysid_and_publish,
+    run_sysid_and_print,
+    live_sysid_and_publish,
 )
-from examples.clustering import (
-    run_clustering_with_local_sysid,
-    run_clustering_with_remote_sysid,
-    run_live_clustering_with_remote_sysid,
+from examples.run_mode_clustering import (
+    run_mode_clustering_with_local_sysid,
+    run_mode_clustering_with_remote_sysid,
+    run_live_mode_clustering_with_remote_sysid,
 )
-from examples.mode_tracking import (
+from examples.run_mode_tracking import (
     run_mode_tracking_with_local_sysid,
     run_mode_tracking_with_remote_sysid,
     run_live_mode_tracking_with_remote_sysid,
 )
-from examples.updating_parameters import (
+from examples.run_model_update import (
     run_model_update_local_sysid, 
     run_model_update_remote_sysid
 )
@@ -38,44 +38,49 @@ def accelerometers(ctx):
 
 @cli.command()
 @click.pass_context
+def live_accelerometers(ctx):
+    live_read_accelerometers(ctx.obj["CONFIG"])
+
+@cli.command()
+@click.pass_context
 def align_readings(ctx):
     align_acceleration_readings(ctx.obj["CONFIG"])
 
 
 @cli.command()
 @click.pass_context
-def oma_and_publish(ctx):
-    run_oma_and_publish(ctx.obj["CONFIG"])
+def sysid_and_publish(ctx):
+    run_sysid_and_publish(ctx.obj["CONFIG"])
 
 @cli.command()
 @click.pass_context
-def oma_and_publish_looping(ctx):
-    run_oma_and_publish_loop(ctx.obj["CONFIG"])
+def live_sysid_publish(ctx):
+    live_sysid_and_publish(ctx.obj["CONFIG"])
 
 @cli.command()
 @click.pass_context
-def oma_and_plot(ctx):
-    run_oma_and_plot(ctx.obj["CONFIG"])
+def sysid_and_plot(ctx):
+    run_sysid_and_plot(ctx.obj["CONFIG"])
 
 @cli.command()
 @click.pass_context
-def oma_and_print(ctx):
-    run_oma_and_print(ctx.obj["CONFIG"])
+def sysid_and_print(ctx):
+    run_sysid_and_print(ctx.obj["CONFIG"])
 
 @cli.command()
 @click.pass_context
 def clustering_with_local_sysid(ctx):
-    run_clustering_with_local_sysid(ctx.obj["CONFIG"])
+    run_mode_clustering_with_local_sysid(ctx.obj["CONFIG"])
 
 @cli.command()
 @click.pass_context
 def clustering_with_remote_sysid(ctx):
-    run_clustering_with_remote_sysid(ctx.obj["CONFIG"])
+    run_mode_clustering_with_remote_sysid(ctx.obj["CONFIG"])
 
 @cli.command()
 @click.pass_context
 def live_clustering_with_remote_sysid(ctx):
-    run_live_clustering_with_remote_sysid(ctx.obj["CONFIG"])
+    run_live_mode_clustering_with_remote_sysid(ctx.obj["CONFIG"])
 
 @cli.command()
 @click.pass_context
