+++ src/examples/run_model_update.py
@@ -1,9 +1,9 @@
 import time
 from data.comm.mqtt import load_config
 from data.accel.hbk.aligner import Aligner
-from methods import sysid_module as sysID
-from methods import clustering_tracking_module as MT
-from methods import model_update_module as MU
+from methods import sysid as sysID
+from methods import mode_clustering as MC
+from methods import model_update as MU
 from methods.constants import PARAMS
 # pylint: disable=R0914, C0103
 
@@ -24,11 +24,11 @@ def run_model_update_local_sysid(config_path):
     while aligner_time is None:
         print("Not enough aligned yet")
         time.sleep(10)
-        oma_output, aligner_time = sysID.get_oma_results(number_of_minutes, aligner, fs)
+        oma_output, aligner_time = sysID.get_sysid_results(number_of_minutes, aligner, fs)
     data_client.disconnect()
 
     # Mode Track
-    dictionary_clusters, median_frequencies = MT.run_mode_clustering(oma_output,PARAMS)
+    dictionary_clusters, median_frequencies = MC.cluster_sysid(oma_output,PARAMS)
 
     # Run model update
     update_result = MU.run_model_update(dictionary_clusters)
@@ -60,12 +60,12 @@ def run_model_update_local_sysid(config_path):
 
 def run_model_update_remote_sysid(config_path):
     config = load_config(config_path)
-    cleaned_values, _, _ = (
-        MT.subscribe_and_get_cleaned_values(config_path)
+    sysid_output, clusters, median_frequencies = (
+        MC.subscribe_and_cluster(config_path)
     )
 
     # Run model update
-    update_result = MT.run_model_update(cleaned_values)
+    update_result = MU.run_model_update(clusters)
 
     if update_result is not None:
         optimized_parameters = update_result['optimized_parameters']
