+++ src/examples/run_sysid.py
@@ -3,14 +3,14 @@ import time
 import matplotlib.pyplot as plt
 from data.comm.mqtt import load_config
 from data.accel.hbk.aligner import Aligner
-from functions.sysid_plot import plot_stabilization_diagram
-from methods import sysid_module as sysID
+from functions.plot_sysid import (plot_stabilization_diagram, plot_pre_stabilization_diagram)
+from methods import sysid as sysID
 from methods.constants import PARAMS
 
 
-def setup_oma(config_path, data_topic_indexes):
+def setup_sysid(config_path, data_topic_indexes):
     """
-    Helper function to set up OMA (Operational Modal Analysis).
+    Helper function to set up sysid (Operational Modal Analysis).
 
     Parameters:
         config_path (str): Path to the configuration file.
@@ -32,11 +32,12 @@ def setup_oma(config_path, data_topic_indexes):
     return aligner, data_client, fs
 
 
-def run_oma_and_plot(config_path):
+def run_sysid_and_plot(config_path):
     number_of_minutes = 1
     data_topic_indexes = [0, 2, 3, 4]
-    aligner, data_client, fs = setup_oma(config_path, data_topic_indexes)
+    aligner, data_client, fs = setup_sysid(config_path, data_topic_indexes)
 
+    fig_ax1 = None
     fig_ax = None
     aligner_time = None
     t1 = time.time()
@@ -45,18 +46,19 @@ def run_oma_and_plot(config_path):
         t2 = time.time()
         t_text = f"Waiting for data for {round(t2-t1,1)} seconds"
         print(t_text,end="\r")
-        results, aligner_time = sysID.get_oma_results(number_of_minutes, aligner, fs)
+        results, aligner_time = sysID.get_sysid_results(number_of_minutes, aligner, fs)
     data_client.disconnect()
     print(aligner_time)
+    fig_ax1 = plot_pre_stabilization_diagram(results, PARAMS, fig_ax=fig_ax1)
     fig_ax = plot_stabilization_diagram(results, PARAMS, fig_ax=fig_ax)
     plt.show(block=True)
     sys.stdout.flush()
 
 
-def run_oma_and_print(config_path):
+def run_sysid_and_print(config_path):
     number_of_minutes = 0.2
     data_topic_indexes = [0, 2, 3, 4]
-    aligner, data_client, fs = setup_oma(config_path, data_topic_indexes)
+    aligner, data_client, fs = setup_sysid(config_path, data_topic_indexes)
 
     aligner_time = None
     t1 = time.time()
@@ -65,7 +67,7 @@ def run_oma_and_print(config_path):
         t2 = time.time()
         t_text = f"Waiting for data for {round(t2-t1,1)} seconds"
         print(t_text,end="\r")
-        results, aligner_time = sysID.get_oma_results(number_of_minutes, aligner, fs)
+        results, aligner_time = sysID.get_sysid_results(number_of_minutes, aligner, fs)
     data_client.disconnect()
     sys.stdout.flush()
 
@@ -75,16 +77,16 @@ def run_oma_and_print(config_path):
     print(f"\n cov_damping \n{results['Xi_poles_cov']}")
 
 
-def run_oma_and_publish(config_path):
+def run_sysid_and_publish(config_path):
     number_of_minutes = 1
     data_topic_indexes = [0, 2, 3, 4]
-    aligner, data_client, fs = setup_oma(config_path, data_topic_indexes)
+    aligner, data_client, fs = setup_sysid(config_path, data_topic_indexes)
     publish_config = load_config(config_path)["sysID"]
 
-    # Setting up the client for publishing OMA results
+    # Setting up the client for publishing sysid results
     publish_client, _ = sysID.setup_client(publish_config)  # fs not needed here
 
-    publish_result = sysID.publish_oma_results(
+    publish_result = sysID.publish_sysid_results(
         number_of_minutes,
         aligner,
         publish_client,
@@ -98,23 +100,23 @@ def run_oma_and_publish(config_path):
     sys.stdout.flush()
 
 
-def run_oma_and_publish_loop(config_path):
+def live_sysid_and_publish(config_path):
     number_of_minutes = 1
     data_topic_indexes = [0, 2, 3, 4]
-    aligner, data_client, fs = setup_oma(config_path, data_topic_indexes)
+    aligner, data_client, fs = setup_sysid(config_path, data_topic_indexes)
     publish_config = load_config(config_path)["sysID"]
 
-    # Setting up the client for publishing OMA results
+    # Setting up the client for publishing sysid results
     publish_client, _ = sysID.setup_client(publish_config)  # fs not needed here
 
-    loop = True
-    while loop:
-        loop = sysID.publish_oma_results(
+    publish_result = True
+    while publish_result:
+        publish_result = sysID.publish_sysid_results(
             number_of_minutes,
             aligner,
             publish_client,
             publish_config["TopicsToSubscribe"][0],
             fs
         )
-        if loop is True:
+        if publish_result is True:
             print(f"Publishing to topic: {publish_config['TopicsToSubscribe'][0]}")
